refactor: delete legacy buy/onramp screens and viewmodels
The amount-first buy flow replaces them — drop CurrencyBuyAmountScreen,
CurrencyBuyViewModel, OnrampAmountScreen, and OnrampViewModel along with the
sheet plumbing in CurrencyInfoScreen. Extract OnrampDeeplinkInbox,
OnrampVerificationPath, and ApplePayOverlay into their own files alongside
OnrampCoordinator, and migrate the buy regression tests to BuyAmountViewModel.
FundingSelectionSheet stays; the launch wizard still uses it.
