+++ Flipcash/Core/Controllers/Onramp/OnrampCoordinator.swift
@@ -922,6 +922,10 @@ final class OnrampCoordinator {
 
 // MARK: - Supporting types -
 
+enum OnrampError: Error {
+    case missingCoinbaseApiKey
+}
+
 private enum Origin: Int {
     case root
     case info
