--- Flipcash/Core/Screens/Main/Currency Swap/CurrencyBuyAmountScreen.swift
@@ -1,70 +1,0 @@
-//
-//  CurrencyBuyAmountScreen.swift
-//  Code
-//
-//  Created by Raul Riera on 2025-12-18.
-//
-
-import SwiftUI
-import FlipcashUI
-import FlipcashCore
-
-struct CurrencyBuyAmountScreen: View {
-    @Bindable private var viewModel: CurrencyBuyViewModel
-    @Environment(\.dismiss) var dismissAction: DismissAction
-    @Environment(RatesController.self) private var ratesController
-
-    @State private var isShowingCurrencySelection: Bool = false
-
-    // MARK: - Init -
-
-    init(viewModel: CurrencyBuyViewModel) {
-        self.viewModel = viewModel
-    }
-
-    // MARK: - Body -
-
-    var body: some View {
-        NavigationStack(path: $viewModel.path) {
-            Background(color: .backgroundMain) {
-                EnterAmountView(
-                    mode: .buy,
-                    enteredAmount: $viewModel.enteredAmount,
-                    subtitle: .balanceWithLimit(viewModel.maxPossibleAmount),
-                    actionState: $viewModel.actionButtonState,
-                    actionEnabled: { _ in
-                        viewModel.canPerformAction
-                    },
-                    action: viewModel.amountEnteredAction,
-                    currencySelectionAction: showCurrencySelection
-                )
-                .foregroundStyle(.textMain)
-                .padding(20)
-            }
-            .navigationTitle(viewModel.screenTitle)
-            .navigationBarTitleDisplayMode(.inline)
-            .navigationDestination(for: CurrencyBuyPath.self) { step in
-                switch step {
-                case .processing(let swapId, let currencyName, let amount):
-                    SwapProcessingScreen(swapId: swapId, swapType: .buyWithReserves, currencyName: currencyName, amount: amount)
-                        .environment(\.dismissParentContainer, {
-                            dismissAction()
-                        })
-                }
-            }
-            .toolbar {
-                ToolbarItem(placement: .topBarTrailing) {
-                    CloseButton { dismissAction() }
-                }
-            }
-            .dialog(item: $viewModel.dialogItem)
-            .sheet(isPresented: $isShowingCurrencySelection) {
-                CurrencySelectionScreen(ratesController: ratesController)
-            }
-        }
-    }
-
-    private func showCurrencySelection() {
-        isShowingCurrencySelection.toggle()
-    }
-}
